feature add cart and checkout

+++ app/schemas/create_book.py
@@ -1,10 +1,11 @@
 from pydantic import BaseModel, Field
+from decimal import Decimal
 
 
 class CreateBook(BaseModel):
     title: str
     author: str
-    price: float = Field(..., gt=0, description="Price must positive")
+    price: Decimal = Field(..., gt=0, description="Price must positive")
     stock: int = Field(..., ge=0, )
 
     model_config = {
